Deleted everything except the main.dart as we spoke about.
We want to clean up the code by changing the design of the application from a drawer menu to bottom navigation. We tried implementing it but it gave a couple of error which would take a lot of time to resolve. 
That is not even the main issue. The main is issue is that the app doesn't run unless it is hot reloaded 3 times. That means, when is for example in a production environment, it will automatically fail to run. We simply do not want that.